requirements.txt
# requirements.txt
requests==2.32.3
adafruit-circuitpython-seesaw==1.16.7
gpiozero==2.0
python-dotenv